Broken build; converted two-dimensional array to Math.NET SparseMatrix for simplified LU factorization

+++ src/ColleyMatrix/Model/Team.cs
@@ -5,6 +5,6 @@
         public int TeamId { get; set; }
         public int Wins { get; set; }
         public int Losses { get; set; }
-        public decimal ColleyRating { get; set; }
+        public double ColleyRating { get; set; }
     }
 }
