+++ src/ColleyMatrix/Service/IColleyMatrixService.cs
@@ -4,6 +4,6 @@
     {
         void SimulateGame(int winnerId, int loserId);
         void Solve();
-        decimal ComputeRating(int wins, int losses);
+        double ComputeRating(double wins, double losses);
     }
 }
